refactor: migrate remaining db/queries.rs functions to repositories
* Migrate the remaining legacy database functions from db/queries.rs to
  appropriate repositories and delete the now-empty db/ module.

* Updated handlers to use repositories

* Removed src/db/queries.rs
