+++ src/api/orgs/settings.rs
@@ -3,17 +3,17 @@
 /// GET  /api/orgs/{id}/settings - Get org settings
 /// PATCH /api/orgs/{id}/settings - Update org settings
 use crate::auth;
-use crate::db;
 use crate::models::Tier;
-use crate::repositories::OrgRepository;
+use crate::repositories::{BillingRepository, OrgRepository};
 use crate::utils::AppError;
 use worker::d1::D1Database;
 use worker::*;
 
 /// Helper to get effective tier for an organization
 async fn get_org_tier(db: &D1Database, org: &crate::models::Organization) -> Tier {
+    let billing_repo = BillingRepository::new();
     if let Some(ref billing_account_id) = org.billing_account_id
-        && let Ok(Some(billing_account)) = db::get_billing_account(db, billing_account_id).await
+        && let Ok(Some(billing_account)) = billing_repo.get_by_id(db, billing_account_id).await
     {
         return Tier::from_str_value(&billing_account.tier).unwrap_or(Tier::Free);
     }
