--- src/db/mod.rs
@@ -1,3 +1,0 @@
-pub mod queries;
-
-pub use queries::*;
