+++ src/repositories/billing_repository.rs
@@ -273,6 +273,23 @@ impl BillingRepository {
         Ok(true)
     }
 
+    /// Reset monthly counter for a billing account (admin only, for testing).
+    pub async fn reset_monthly_counter(
+        &self,
+        db: &D1Database,
+        billing_account_id: &str,
+        year_month: &str,
+    ) -> Result<()> {
+        let stmt = db.prepare(
+            "DELETE FROM monthly_counters
+             WHERE billing_account_id = ?1 AND year_month = ?2",
+        );
+        stmt.bind(&[billing_account_id.into(), year_month.into()])?
+            .run()
+            .await?;
+        Ok(())
+    }
+
     /// Count organizations linked to a billing account.
     pub async fn count_orgs(&self, db: &D1Database, billing_account_id: &str) -> Result<i64> {
         let result = db
